Distinguish `negate` flag from `reverse` flag
Assertions that an element is not visible are throwing errors when the
element exists. Similarly, asserting that the count of elements does not
match a given number also throws an error. This is because the `negate`
flag is being treated the same as the `reverse` flag for `waitForExist`,
causing the assertions to blow up when the element exists.

Note that negating an assertion does not necessarily mean the element
should not exist. In the case of `visible`, we could be checking that an
element that exists in the DOM is not visible (ex. visbility hidden).
Likewise, stating that the count of elements is not equal to a number
does not mean that we expect no instance of that element type to exist.

Tests did not catch this since `elementExists` is stubbed throughout, so
the actual `waitForExist` is never called / its behavior is not properly
mimicked.