Use `waitForVisible` instead of `waitForExist` in `visible`
Switches out `waitForExist` for `waitForVisible`, since this is the
expected behavior when working with wait times and asserting whether or
not something is visible (i.e. when a wait is specified, the client will
wait that amount of time for that element to be visible/not visible, as
opposed to waiting for it to exist/not exist).