Use `waitUntil` instead of `waitForExist` in `count`
Switches out `waitForExist` for `waitUntil`, since this is the expected
behavior when working with wait times and asserting whether or not an
element appears an expected amount of times  (i.e. when a wait is
specified, the client will wait that amount of time for that element
to be appear or not appear a certain number of times, as opposed to
waiting for it to exist/not exist).